+++ changelog.md
@@ -1,3 +1,8 @@
+#0.2.4
+* Correction de l'instanciation du Core.Models.Entreprise
+* Création de Core.Models.IEntreprise : permettant de gérer les paramèters d'instanciation de Core.Models.Entreprise
+* MEP des test de Core.Models.Entreprise
+
 #0.1.0
 * Initialisation du projet 
 * Inclusion des travaux précédents : QreaFacturation et QreaCompta